drop registry references
Summary:

rough sweep to get rid of references to registry related code.
bosh-init dropped from gomodules because it's not required anymore.

the registry has been deprecated and removed in the bosh-director. This
removes code related to the registry from the agents codebase to
continue the registry deprecation.

aside from removing the code from the agent, this commit also deals with
some implications of the removal.

- integration tests are not utilizing the registry to setup the agent
  anymore. the registry usage was replaced with fileSettings.

- update fixtures to reflect the removal of the registry and switch them
  over to filesettings